third_party/blink/web_tests/fast/dom/shadow/querySelector-with-detached-node-distribution-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
crbug.com/337618: querySelector needs ensure distribution.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS barShadowRoot.querySelectorAll(":host-context(.foo)::shadow div")[0].id is "bar"
PASS barShadowRoot.querySelectorAll(":host-context(.foo)::shadow div")[0] is undefined
PASS successfullyParsed is true

TEST COMPLETE
